Update GitHub workflows to publish GNU Awk
In addition to converting file reference from `tar` to `gawk`, we also
take the opportunity to use an explicit (but trivial) install script to
place build outputs in desired locations. This is preferred over copying
files out of the build output tree after the fact since the layout of
this tree is not entirely stable (esp. WRT bzlmod).

In addition, we strive to act more against aggregating globs than
individual files.