Promote build settings to common settings
It used to be an error when a flag got inherrited from `common` by a
command that didn't accept that flag, but the modern sementics apply
`common` settings whenever they match but ignore them when they don't.
This makes it easier to keep using the same settings consistently for
all commands.